requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
geopandas==0.14.1
shapely==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0

